feat(zkp): refactored zok code ready for merging
