+++ index.js
@@ -2,7 +2,7 @@ const generateZokratesFiles = require('./setup/generateZokratesFiles');
 const merkleTree = require('./merkleTree');
 const provider = require('./provider');
 const vks = require('./vks');
-const erc20 = require('./erc20');
+const erc20 = require('./erc20rc');
 const erc721 = require('./erc721');
 const utils = require('./utils');
 const elgamal = require('./elgamal');
