+++ setup/gm17/rc/ft-mint.zok
