+++ setup/gm17/rc/nft-burn.zok
